änderungen

+++ src/main/java/fhv/team11/project/ems/booking/transfer/BookingDTO.java
@@ -8,6 +8,7 @@ import fhv.team11.project.ems.events.transfer.ActiveEventWizardDTO;
 import fhv.team11.project.ems.user.transfer.UserDTO;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -18,6 +19,7 @@ public class BookingDTO {
     private AddressDTO participantAddress;
 
     private UserDTO participant;
+    //TODO participant to String
     private UserDTO reservationist;
 
     private ActiveEventListDTO bookedEvent;
@@ -30,6 +32,7 @@ public class BookingDTO {
 
     private LocalDate optionDate;
 
-    private int price;
+    private BigDecimal price;
+
 
 }
